docker update, guidance for the data loading module implementation
