+++ data/loader/resources/__init__.py
@@ -1,0 +1,6 @@
+from pathlib import Path
+
+ROOT = Path(__file__).parent
+
+# Example
+# MY_SUPER_FILE = ROOT.joinpath("my_super_file.txt")
